Split src/bindings.cpp one-file-per-subsystem
The bindings file followed the namespace structure already — every
section was prefixed with `// === <ClassName> ===` and routed to the
appropriate submodule — but it all lived in a single 1500-line file.
This commit pulls each section into its subsystem's own Bindings.cpp:

  src/mesh/Bindings.cpp         — Edge, Vertex, Simplex, EdgeList,
                                   VertexList, SimplexOrientation,
                                   SimplexFilter and friends
  src/spacetime/Bindings.cpp    — Spacetime, Metric, Signature, Foliation,
                                   topologies, Pachner moves
  src/observables/Bindings.cpp  — SparseGraph, ModularityOptimizer,
                                   VolumeProfile, WilsonLoop
  src/simulations/Bindings.cpp  — CDT, ReggeSolver
  src/quantum/Bindings.cpp      — renamed from bindings.cpp; register_quantum
                                   renamed from register_quantum_bindings
  src/bindings.cpp              — orchestrator: PYBIND11_MODULE entry,
                                   submodule definitions, root-namespace
                                   things (HingeType, MatterConfiguration,
                                   forceLayout3D), and the per-subsystem
                                   register_<sub>(m) dispatch

CMakeLists.txt was updated so the new per-subsystem Bindings.cpp files
land in `_tessera` (not in `tessera_core` / `tessera_quantum`), and so
the quantum static lib excludes the renamed `quantum/Bindings.cpp`.

Test surface unchanged: 210 quantum tests pass, 422 core tests pass.
